requirements.txt
fastapi
uvicorn[standard]
PyMuPDF
google-generativeai
